feat: implement start time flow

+++ app/src/main/java/com/rafaeltmbr/stopwatch/domain/entities/Lap.kt
@@ -2,6 +2,6 @@ package com.rafaeltmbr.stopwatch.domain.entities
 
 data class Lap(
     val index: Int,
-    val time: Float,
-    val diff: Float
+    val time: Long, // ms
+    val diff: Long // ms
 )
